+++ test/test_app.py
@@ -1,8 +1,7 @@
 import re
-from unittest.mock import MagicMock
 import pytest
 import requests
-from test.helpers import is_token, relative_dir, DOD_SDN, DOD_SDN_INFO
+from test.helpers import is_token, relative_dir
 from authnid.make_app import make_config
 from authnid.user_repo import UserRepo
 
@@ -34,31 +33,3 @@ def test_log_in_with_cac_redirects_with_token(request_client):
     assert len(location) < 2000 # ensure URL is under a safe limit
     assert response.status_code == 302
     assert is_token(token)
-
-@pytest.fixture
-def dod_user():
-    user_info = DOD_SDN_INFO.copy()
-    return user_info
-
-def test_adds_new_user_to_the_database(monkeypatch, user_repo, client, dod_user):
-    added_user = {}
-    magic = MagicMock(return_value='123')
-    monkeypatch.setattr('authnid.root.is_valid_certificate', lambda r: True)
-    monkeypatch.setattr('authnid.user_repo.UserRepo.add_user', magic)
-    resp = client.get('/', environ_base={
-        'HTTP_X_SSL_CLIENT_VERIFY': 'SUCCESS',
-        'HTTP_X_SSL_CLIENT_S_DN': DOD_SDN
-    })
-    magic.assert_called_with(**dod_user)
-
-
-def test_does_not_add_existing_user(monkeypatch, user_repo, client):
-    monkeypatch.setattr('authnid.root.is_valid_certificate', lambda r: True)
-    user_info = DOD_SDN_INFO.copy()
-    user_repo.add_user(**user_info)
-    user_count = user_repo.count()
-    resp = client.get('/', environ_base={
-        'HTTP_X_SSL_CLIENT_VERIFY': 'SUCCESS',
-        'HTTP_X_SSL_CLIENT_S_DN': DOD_SDN
-    })
-    assert user_repo.count() == user_count
